% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{alleleFreqMatGeneration}
\alias{alleleFreqMatGeneration}
\title{Generation of the matrix for calculating allelic frequencies}
\usage{
alleleFreqMatGeneration(genomeObj)
}
\arguments{
\item{genomeObj}{a \code{Genome} object}
}
\value{
A matrix for calculating allelic frequencies from genotypes
frequencies.
}
\description{
Generates a matrix that allows to go from genotypic frequencies
to allelic frequencies.
}
\details{
An allele frequency matrix is a matrix with rows equal to the number of
genotypes and columns equal to the number of alleles. By multiplying a
row matrix of genotype frequencies we obtain a row matrix of associated
allele frequencies.
}
\author{
Ehouarn Le Faou
}
