% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{areThereHomoz}
\alias{areThereHomoz}
\title{Are there any allelic combinations including homozygosity}
\usage{
areThereHomoz(formula)
}
\arguments{
\item{formula}{a selection formula}
}
\value{
logical indicating if there are homozygotes
}
\description{
Test if there are homozygotes in the specified allelic combinations of
a selection formula
}
\author{
Ehouarn Le Faou
}
