% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{catn}
\alias{catn}
\title{Concatenate, print and line break}
\usage{
catn(..., file = "", sep = " ", fill = FALSE, labels = NULL, append = FALSE)
}
\arguments{
\item{...}{see \link[base]{cat}.}

\item{file}{see \link[base]{cat}.}

\item{sep}{see \link[base]{cat}.}

\item{fill}{see \link[base]{cat}.}

\item{labels}{see \link[base]{cat}.}

\item{append}{see \link[base]{cat}.}
}
\value{
None (invisible \code{NULL}).
}
\description{
Object output in the same way as the function \link[base]{cat} but adding
a line break at the end.
}
\details{
See \link[base]{cat}.
}
\author{
Ehouarn Le Faou
}
