% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{changeInitGenoFreq}
\alias{changeInitGenoFreq}
\title{Changing the initial genotype frequencies of a population}
\usage{
changeInitGenoFreq(population, initGenoFreq)
}
\arguments{
\item{initGenoFreq}{a vector of the size of the genotype number
describing the allele frequencies common to all simulations}

\item{populations}{a list of \code{Population} objects}
}
\value{
a \code{Population} object with the new initial genotype frequencies
}
\description{
Changing the initial genotype frequencies of a population
}
\examples{
# Definition of a population in its simplest form:
DL <- list(dl = c("A", "a"))
HL <- list(hl = c("B", "b"))
mutations <- list(
  mutation(from = "a", to = "A", rate = 1e-3),
  mutation(from = "b", to = "B", rate = 1e-3)
)
genomeObj <- setGenome(listHapLoci = HL, listDipLoci = DL)
pop <- setPopulation(
  name = "A",
  size = 1000,
  dioecy = TRUE,
  genomeObj = genomeObj,
  selectionObj = setSelectNeutral(genomeObj),
  mutMatrixObj = setMutationMatrix(genomeObj, mutations = mutations)
)
pop = changeinitGenoFreq(pop, c(0, 0, 0, 0, 0, 0, 0, 0, 0, 1))

}
\author{
Ehouarn Le Faou
}
