% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{check.ease}
\alias{check.ease}
\title{The validity check for the \code{Ease} class}
\usage{
check.ease(object)
}
\arguments{
\item{object}{\code{Ease} object}
}
\value{
A logical corresponding to whether the object is a correct
\code{Ease} object.
}
\description{
The validity check for the \code{Ease} class
}
\author{
Ehouarn Le Faou
}
