% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeClass.R
\name{check.genome}
\alias{check.genome}
\title{The validity check for the \code{Genome} class}
\usage{
check.genome(object)
}
\arguments{
\item{object}{a \code{Genome} object}
}
\value{
A logical corresponding to whether the object is a correct
\code{Genome} object.
}
\description{
The validity check for the \code{Genome} class
}
\author{
Ehouarn Le Faou
}
