% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetapopulationClass.R
\name{check.metapopulation}
\alias{check.metapopulation}
\title{The validity check for the \code{Metapopulation} class}
\usage{
check.metapopulation(object)
}
\arguments{
\item{object}{a \code{Metapopulation} object}
}
\value{
a boolean corresponding to whether the object is a correct
\code{Metapopulation} object.
}
\description{
The validity check for the \code{Metapopulation} class
}
\author{
Ehouarn Le Faou
}
