% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationMatrixClass.R
\name{check.mutationMatrix}
\alias{check.mutationMatrix}
\title{The validity check associated with the \code{MutationMatrix} class}
\usage{
check.mutationMatrix(object)
}
\arguments{
\item{object}{an object of class \code{MutationMatrix}}
}
\value{
A logical corresponding to whether \code{x} is a correct
\code{MutationMatrix} object.
}
\description{
The validity check associated with the \code{MutationMatrix} class
}
\author{
Ehouarn Le Faou
}
