% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationClass.R
\name{check.population}
\alias{check.population}
\title{The validity check for the \code{Population} class}
\usage{
check.population(object)
}
\arguments{
\item{object}{a \code{Population} object}
}
\value{
a boolean corresponding to whether the object is a correct
\code{Population} object.
}
\description{
The validity check for the \code{Population} class
}
\author{
Ehouarn Le Faou
}
