% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionClass.R
\name{check.selection}
\alias{check.selection}
\title{The validity check for the \code{Selection} class}
\usage{
check.selection(object)
}
\arguments{
\item{object}{\code{Selection} object}
}
\value{
A logical corresponding to whether the object is a correct
\code{Selection} object.
}
\description{
The validity check for the \code{Selection} class
}
\author{
Ehouarn Le Faou
}
