% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{genotyping}
\alias{genotyping}
\title{Genotyping}
\usage{
genotyping(genomeObj)
}
\arguments{
\item{genomeObj}{a \code{Genome} object}
}
\value{
A list of matrices describing genotypes in rows.
}
\description{
Generation of genotypes associated with a \code{Genome} object.
}
\details{
The output genotypes are described as a list of three matrices. A
genotype consists of two diploid haplotypes (first two matrices)
and one haploid haplotype (third matrix), which are read at the
same row number on all three matrices.
}
\author{
Ehouarn Le Faou
}
