% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{getCustomOutput}
\alias{getCustomOutput}
\title{Getting the custom output}
\usage{
getCustomOutput(metapop)
}
\arguments{
\item{metapop}{a \code{Metapopulation} objects}
}
\value{
The list generated through the custom result function, if at least
it was specified during the simulation of the \code{Metapopulation}.
}
\description{
Getting the custom output
}
\author{
Ehouarn Le Faou
}
