% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{getResults}
\alias{getResults}
\title{Getting the simulation results}
\usage{
getResults(metapop)
}
\arguments{
\item{metapop}{a \code{Metapopulation} objects}
}
\value{
A data.frame where each line corresponds to a simulation. The
results include :
- the last generation reached (the threshold or the generation that first
verified at least one of the stopping conditions)
- the final population size
- the genotype frequencies
- allelic frequencies
- the reason(s) for the stop (either the threshold was reached, i.e.
\code{unstopped} or the stop condition(s) that was (were) reached, in the
form of boolean values
- Average fitness (individual, gamete production and gametic)
}
\description{
Getting the simulation results
}
\author{
Ehouarn Le Faou
}
