% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{haplotyping}
\alias{haplotyping}
\title{Haplotyping}
\usage{
haplotyping(genomeObj)
}
\arguments{
\item{genomeObj}{a \code{Genome} object}
}
\value{
A list of matrices describing haplotypes in rows.
}
\description{
Generation of haplotypes associated with a \code{Genome} object.
}
\details{
The generated haplotypes are output as a list of three enumeration
in the form of matrices of alleles (each row corresponding to an haplotype,
each column to a locus). The first enumeration corresponds to haplotypes
considering only haploid loci, the second only diploid loci and the third
all loci (with two matrices, 1 for haploid loci, 1 for diploid loci).
}
\author{
Ehouarn Le Faou
}
