% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{IDgenomeGeneration}
\alias{IDgenomeGeneration}
\title{Genome identifier}
\usage{
IDgenomeGeneration(listLoci, alleles)
}
\arguments{
\item{listLoci}{the list of all loci}

\item{alleles}{the vector of all the alleles}
}
\value{
The genome ID as a character string.
}
\description{
Generation of the input genome ID, i.e. the concatenation in string form
of the names of the loci and alleles constituting this genome.
}
\author{
Ehouarn Le Faou
}
