% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{IDgenotypeGeneration}
\alias{IDgenotypeGeneration}
\title{Genotype identifier}
\usage{
IDgenotypeGeneration(dl1, dl2, hl = NULL)
}
\arguments{
\item{dl1}{the first diploid haplotype as a character (or factors) vector.}

\item{dl2}{the second diploid haplotype as a character (or factors) vector.}

\item{hl}{the haploid haplotype as a character (or factors) vector.}
}
\value{
The genotype ID as a character string.
}
\description{
Generation of the input genotype ID, i.e. the concatenation in string form
of the names of the alleles constituting these haplotypes (the two from
the diploid genome and the one from the haploid genome).
}
\author{
Ehouarn Le Faou
}
