% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{IDhaplotypeGeneration}
\alias{IDhaplotypeGeneration}
\title{Haplotype identifier}
\usage{
IDhaplotypeGeneration(dl, hl)
}
\arguments{
\item{dl}{character (or factors) vector. The diploid haplotype}

\item{hl}{character (or factors) vector. The haploid haplotype}
}
\value{
The haplotype ID as a character string.
}
\description{
Generation of the input haplotype ID, i.e. the concatenation in string form
of the names of the alleles constituting this haplotype.
}
\author{
Ehouarn Le Faou
}
