% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{initialize,Ease-method}
\alias{initialize,Ease-method}
\title{Initialize method for the \code{Ease} class}
\usage{
\S4method{initialize}{Ease}(
  .Object,
  N,
  threshold,
  dioecy,
  mutMatrixObj,
  genomeObj,
  selectionObj,
  stopCondition,
  selfRate,
  initGenoFreq
)
}
\arguments{
\item{.Object}{a \code{Ease} object}

\item{N}{the population size}

\item{threshold}{the maximum number of generations}

\item{dioecy}{logical indicating whether the simulated population is
dioecious or hermaphroditic}

\item{mutMatrixObj}{a \code{MutationMatrix} object}

\item{genomeObj}{a \code{Genome} object}

\item{selectionObj}{a \code{Selection} object}

\item{stopCondition}{list of vectors that each describe the alleles that
must be fixed to define a stop condition. Each of these stop conditions
will therefore be associated with a stop condition}

\item{selfRate}{the selfing rate}

\item{initGenoFreq}{A vector of the size of the genotype number
describing the initial allele frequencies common to all simulations}
}
\value{
An \code{Ease} object
}
\description{
Initialize method for the \code{Ease} class
}
\author{
Ehouarn Le Faou
}
