% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeClass.R
\name{initialize,Genome-method}
\alias{initialize,Genome-method}
\title{Initialize method for the \code{Genome} class}
\usage{
\S4method{initialize}{Genome}(.Object, listHapLoci, listDipLoci, recRate)
}
\arguments{
\item{.Object}{a \code{Genome} object}

\item{listHapLoci}{a list of haploid loci}

\item{listDipLoci}{a list of diploid loci}

\item{recRate}{a two-by-two recombination rate vector}
}
\value{
A \code{Genome} object
}
\description{
Initialize method for the \code{Genome} class
}
\author{
Ehouarn Le Faou
}
