% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetapopulationClass.R
\name{initialize,Metapopulation-method}
\alias{initialize,Metapopulation-method}
\title{Initialize method for the \code{Metapopulation} class}
\usage{
\S4method{initialize}{Metapopulation}(.Object, populations, migMat)
}
\arguments{
\item{.Object}{a \code{Metapopulation} object}

\item{populations}{list of \code{Population} object(s)}

\item{migMat}{migration matrix}
}
\value{
a \code{Metapopulation} object
}
\description{
Initialize method for the \code{Metapopulation} class
}
\author{
Ehouarn Le Faou
}
