% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationMatrixClass.R
\name{initialize,MutationMatrix-method}
\alias{initialize,MutationMatrix-method}
\title{Initialize method for the \code{MutationMatrix} class}
\usage{
\S4method{initialize}{MutationMatrix}(.Object, genomeObj, mutHapLoci, mutDipLoci)
}
\arguments{
\item{.Object}{a \code{MutationMatrix} object}

\item{genomeObj}{a \code{Genome} object}

\item{mutHapLoci}{a list of haploid locus by locus allelic mulation matrices.}

\item{mutDipLoci}{a list of diploid locus by locus allelic mulation matrices.}
}
\value{
A \code{MutationMatrix} object
}
\description{
Initialize method for the \code{MutationMatrix} class
}
\author{
Ehouarn Le Faou
}
