% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionClass.R
\name{initialize,Selection-method}
\alias{initialize,Selection-method}
\title{Initialize method for the \code{Selection} class}
\usage{
\S4method{initialize}{Selection}(.Object, genomeObj)
}
\arguments{
\item{.Object}{a \code{Selection} object}

\item{genomeObj}{a \code{Genome} object}
}
\value{
A \code{Selection} object
}
\description{
Initialize method for the \code{Selection} class
}
\author{
Ehouarn Le Faou
}
