% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationMatrixClass.R
\name{is.correct.mut.matrix}
\alias{is.correct.mut.matrix}
\title{Test if a matrix is a correct mutation matrix}
\usage{
is.correct.mut.matrix(x, name)
}
\arguments{
\item{x}{a matrix.}

\item{name}{the name of the matrix}
}
\value{
A logical corresponding to whether \code{x} is a correct mutation
matrix, i.e. a square matrix with dimensions greater than 0 and whose rows
sum to 1.
}
\description{
Test if a matrix is a correct mutation matrix
}
\author{
Ehouarn Le Faou
}
