% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{is.correct.transition.matrix}
\alias{is.correct.transition.matrix}
\title{Test if a matrix is a correct transition matrix}
\usage{
is.correct.transition.matrix(x, type, name)
}
\arguments{
\item{x}{a matrix.}

\item{type}{type of the matrice (mutation matrix ? recombination matrix ?)}

\item{name}{the name of the matrix.}
}
\value{
A logical corresponding to whether \code{x} is a correct transition
matrix, i.e. a square matrix with dimensions greater than 0 and whose rows
sum to 1.
}
\description{
Test if a matrix is a correct transition matrix
}
\author{
Ehouarn Le Faou
}
