% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{is.default.matrix}
\alias{is.default.matrix}
\title{Test if a matrix is a default matrix}
\usage{
is.default.matrix(x)
}
\arguments{
\item{x}{a matrix.}
}
\value{
A logical corresponding to whether \code{x} is a default matrix
(matrix of dimension 0x0).
}
\description{
Test if a matrix is a default matrix
}
\author{
Ehouarn Le Faou
}
