% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{is.probability.matrix}
\alias{is.probability.matrix}
\title{Test if a matrix is of probability}
\usage{
is.probability.matrix(x)
}
\arguments{
\item{x}{a matrix.}
}
\value{
A logical corresponding to whether \code{x} is a probability
matrix (sum of rows equal to 1).
}
\description{
Test if a matrix is of probability
}
\author{
Ehouarn Le Faou
}
