% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{isAffected}
\alias{isAffected}
\title{Is this haplo/geno-type affected ?}
\usage{
isAffected(refDNAtype, selDNAtype)
}
\arguments{
\item{refDNAtype}{the reference allelic combination (that of a genotype or a
haplotype}

\item{selDNAtype}{the selected allelic combination}
}
\value{
a logic indicating whether the reference genotype or haplotype
is affected by the allelic combination under selection
}
\description{
Determination for a given genotype or haplotype whether it
contains the allelic combination under selection
}
\author{
Ehouarn Le Faou
}
