% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{listing}
\alias{listing}
\title{Listing for display}
\usage{
listing(vect)
}
\arguments{
\item{vect}{a vector of any class.}
}
\value{
A listing of the elements of the input vector as a string.
}
\description{
Listing from the elements of a vector by producing a string, with comma
separation between each element and the word "and" between the last two
elements.
}
\author{
Ehouarn Le Faou
}
