% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MATRIX_PRODUCT}
\alias{MATRIX_PRODUCT}
\title{Matrix product}
\usage{
MATRIX_PRODUCT(MAT1, MAT2)
}
\arguments{
\item{MAT1}{matrix}

\item{MAT2}{matrix}
}
\description{
Standard matrix product. The first matrix must have as many columns as the
second has rows.
}
\author{
Ehouarn Le Faou
}
