% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{mutMatFriendly}
\alias{mutMatFriendly}
\title{Individual mutation definition to allelic mutation matrices}
\usage{
mutMatFriendly(genomeObj, mutations)
}
\arguments{
\item{genomeObj}{a \code{Genome} object}

\item{mutations}{list of mutations defined individually with the function
\link[Ease]{mutation}}
}
\value{
A list of the two list of allelic mutation matrices, for haploid
and diploid loci respectively.
}
\description{
Translation of the list of individually defined mutations into allelic
mutation matrices which are then used to generate the genotypic mutation
matrix.
}
\author{
Ehouarn Le Faou
}
