% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{mutMatRates}
\alias{mutMatRates}
\title{Mutation matrix from rates}
\usage{
mutMatRates(alleles, forwardMut, backwardMut)
}
\arguments{
\item{alleles}{allele enumeration vector of a locus}

\item{forwardMut}{forward mutation rate}

\item{backwardMut}{backward mutation rate}
}
\value{
An allelic mutation matrix (probability matrix which associates to
each allele in a row the probability of mutating or not to the other alleles
of the locus in question).
}
\description{
Generation of a mutation matrix from the allele enumeration vector of a loci
and the forward and backward mutation rates.
}
\details{
See \code{MutationMatrix} for more details on mutation matrices.
}
\author{
Ehouarn Le Faou
}
