% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{plot,Ease-method}
\alias{plot,Ease-method}
\title{Plot method for the \code{Ease} class}
\usage{
\S4method{plot}{Ease}(x, y, ..., n = NULL)
}
\arguments{
\item{x}{a \code{Ease} object}

\item{y}{Ignored.}

\item{...}{Ignored.}

\item{n}{number of the graph to be displayed (between 1 and 3 if no
simulations were recorded, from 1 to 4 otherwise).}
}
\value{
The graphics are displayed one after the other and require user
interaction to switch between them. Three graphs concern the results of
the simulations: (i) the proportion of stop conditions reached, (ii)
the distribution of simulation durations according to the stop conditions
reached, and finally (iii) the average time to reach stop conditions. If the
simulations were recorded (parameter \code{recording} of the method
\code{simulate} for the class {Ease}), an additional graph is available,
and represents the evolution of allelic frequencies as a function of
generations for each of the simulations. For each simulation, the vector
of generations is standardised (by dividing it by its final value) in
order to have an overview of the ways in which the simulations ended.
}
\description{
Generation of a some simple plots that summaries the simulation
results.
}
\author{
Ehouarn Le Faou
}
