% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{print,Ease-method}
\alias{print,Ease-method}
\title{Print method for the \code{Ease} class}
\usage{
\S4method{print}{Ease}(x, ...)
}
\arguments{
\item{x}{a \code{Ease} object}

\item{...}{there are no more parameters.}
}
\value{
No return value, only a display.
}
\description{
Print method for the \code{Ease} class
}
\author{
Ehouarn Le Faou
}
