% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeClass.R
\name{print,Genome-method}
\alias{print,Genome-method}
\title{Print method for the \code{Genome} class}
\usage{
\S4method{print}{Genome}(x, ...)
}
\arguments{
\item{x}{a \code{Genome} object}

\item{...}{Ignored.}
}
\value{
No return value, only a display.
}
\description{
Print method for the \code{Genome} class
}
\author{
Ehouarn Le Faou
}
