% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetapopulationClass.R
\name{print,Metapopulation-method}
\alias{print,Metapopulation-method}
\title{Print method for the \code{Metapopulation} class}
\usage{
\S4method{print}{Metapopulation}(x, ...)
}
\arguments{
\item{x}{a \code{Metapopulation} object}

\item{...}{Ignored.}
}
\description{
Print method for the \code{Metapopulation} class
}
\author{
Ehouarn Le Faou
}
