% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationMatrixClass.R
\name{print,MutationMatrix-method}
\alias{print,MutationMatrix-method}
\title{Print method for the \code{MutationMatrix} class}
\usage{
\S4method{print}{MutationMatrix}(x, ...)
}
\arguments{
\item{x}{a \code{MutationMatrix} object}

\item{...}{there are no more parameters.}
}
\value{
No return value, only a display.
}
\description{
Print method for the \code{MutationMatrix} class
}
\author{
Ehouarn Le Faou
}
