% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationClass.R
\name{print,Population-method}
\alias{print,Population-method}
\title{Print method for the \code{Population} class}
\usage{
\S4method{print}{Population}(x, ...)
}
\arguments{
\item{x}{a \code{Population} object}

\item{...}{the other parameter is \code{frame}, which is a logic indicating
whether the frame surrounding the display of the population characteristics
should be displayed or not.}
}
\description{
Print method for the \code{Population} class
}
\author{
Ehouarn Le Faou
}
