% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionClass.R
\name{print,Selection-method}
\alias{print,Selection-method}
\title{Print method for the \code{Selection} class}
\usage{
\S4method{print}{Selection}(x, ...)
}
\arguments{
\item{x}{a \code{Selection} object}

\item{...}{there are no more parameters.}
}
\value{
No return value, only a display.
}
\description{
Print method for the \code{Selection} class
}
\author{
Ehouarn Le Faou
}
