% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{rowResultGen}
\alias{rowResultGen}
\title{Processing a result (or record) list}
\usage{
rowResultGen(x)
}
\arguments{
\item{x}{list of result or record}
}
\value{
Merges the column names of the matrices making up the list with
the names of the matrices, then merges the matrices together.
}
\description{
Processing a result (or record) list
}
\author{
Ehouarn Le Faou
}
