% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{selectInputTreatment}
\alias{selectInputTreatment}
\title{Treatment of selection formulas}
\usage{
selectInputTreatment(selectInput, genomeObj, haplo = FALSE)
}
\arguments{
\item{selectInput}{a selection input}

\item{genomeObj}{a \code{Genome} object}

\item{haplo}{logical indicating whether the selection should apply to
haplotypes (in the case of gametic selection for example)}
}
\value{
a vector of fitness values
}
\description{
Determines whether an entry for the selection is a list of selection
formulas or just a vector. If it is a list of formulas, turns them
into a vector. If it is a vector, does nothing.
}
\author{
Ehouarn Le Faou
}
