% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{selection.form.treatment}
\alias{selection.form.treatment}
\title{Treatment of a selection formula}
\usage{
selection.form.treatment(factors, genomeObj = NULL, checking = FALSE)
}
\arguments{
\item{factors}{formula factors (right-hand members)}

\item{genomeObj}{a \code{Genome} object for the test (see \code{checking}
parameter)}

\item{checking}{logical indicating whether a test verifying the
compatibility of input factors with the genome}
}
\value{
A list of vectors enumerating the allelic combinations that
correspond to the factors
}
\description{
Conversion of the factors of a selection formula into the list of
corresponding allelic combinations
}
\author{
Ehouarn Le Faou
}
