% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{show,Ease-method}
\alias{show,Ease-method}
\title{Show method for the \code{Ease} class}
\usage{
\S4method{show}{Ease}(object)
}
\arguments{
\item{object}{a \code{Ease} object}
}
\value{
No return value, only a display.
}
\description{
Show method for the \code{Ease} class
}
\author{
Ehouarn Le Faou
}
