% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeClass.R
\name{show,Genome-method}
\alias{show,Genome-method}
\title{Show method for the \code{Genome} class}
\usage{
\S4method{show}{Genome}(object)
}
\arguments{
\item{object}{a \code{Genome} object}
}
\value{
No return value, only a display.
}
\description{
Show method for the \code{Genome} class
}
\author{
Ehouarn Le Faou
}
