% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetapopulationClass.R
\name{show,Metapopulation-method}
\alias{show,Metapopulation-method}
\title{Show method for the \code{Metapopulation} class}
\usage{
\S4method{show}{Metapopulation}(object)
}
\arguments{
\item{object}{a \code{Metapopulation} object}
}
\description{
Show method for the \code{Metapopulation} class
}
\author{
Ehouarn Le Faou
}
