% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationMatrixClass.R
\name{show,MutationMatrix-method}
\alias{show,MutationMatrix-method}
\title{Show method for the \code{MutationMatrix} class}
\usage{
\S4method{show}{MutationMatrix}(object)
}
\arguments{
\item{object}{a \code{MutationMatrix} object}
}
\value{
No return value, only a display.
}
\description{
Show method for the \code{MutationMatrix} class
}
\author{
Ehouarn Le Faou
}
