% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionClass.R
\name{show,Selection-method}
\alias{show,Selection-method}
\title{Show method for the \code{Selection} class}
\usage{
\S4method{show}{Selection}(object)
}
\arguments{
\item{object}{a \code{Selection} object}
}
\value{
No return value, only a display.
}
\description{
Show method for the \code{Selection} class
}
\author{
Ehouarn Le Faou
}
