% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{stopGenerations}
\alias{stopGenerations}
\title{Average number of generations to reach stop conditions}
\usage{
stopGenerations(easeObj)
}
\arguments{
\item{easeObj}{an \code{Ease} object}
}
\value{
A data.frame with two columns and as many rows as there are
levels in the vector generated by the stopVect function. The first column
corresponds to the different levels mentioned above and the second to the
verage number of generations to reach stop conditions calculated over all
simulations.
}
\description{
Data frame summarizing for each stop mode of a simulation (stop condition(s)
and reaching the threshold) the average number of generations to reach stop
conditions
}
\author{
Ehouarn Le Faou
}
