% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{stopProportions}
\alias{stopProportions}
\title{Proportion of stop condition reached}
\usage{
stopProportions(easeObj)
}
\arguments{
\item{easeObj}{an \code{Ease} object}
}
\value{
A data.frame with two columns and as many rows as there are
levels in the vector generated by the stopVect function. The first column
corresponds to the different levels mentioned above and the second to the
proportions of all the simulations that led to these types of simulation
stops.
}
\description{
Data frame summarizing for each stop mode of a simulation (stop condition(s)
and reaching the threshold) the proportion of simulations that have reached
them.
}
\author{
Ehouarn Le Faou
}
