% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{stopVect}
\alias{stopVect}
\title{Stop condition reached}
\usage{
stopVect(easeObj, addAllStop = TRUE)
}
\arguments{
\item{easeObj}{an \code{Ease} object}

\item{addAllStop}{logical indicating whether all stop modalities should be
taken into account (if they have not been reached by any of the simulations,
they will be in the levels of the output factor vector)}
}
\value{
A vector of factors
}
\description{
Generation of the vector describing for each simulation which stop
condition(s) have been validated or if the simulation has not been stopped
(unstopped).
}
\author{
Ehouarn Le Faou
}
