% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaseClass.R
\name{summary,Ease-method}
\alias{summary,Ease-method}
\title{Summary method for the \code{Ease} class}
\usage{
\S4method{summary}{Ease}(object, ...)
}
\arguments{
\item{object}{a \code{Ease} object}

\item{...}{Ignored.}
}
\value{
Some statistics summarising the results obtained by simulating:
the proportions of simulations having reached each of the defined stop
conditions or having reached the threshold and the average duration of
the simulations according to these stop conditions.
}
\description{
Generattion of a small summary of the simulation results.
}
