% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{whichHomoz}
\alias{whichHomoz}
\title{Which alleles are homozygous in the input?}
\usage{
whichHomoz(formula)
}
\arguments{
\item{formula}{a selection formula}
}
\value{
the enumeration of alleles that appear at least once homozygous
}
\description{
Determine which alleles are at least once input as homozygous in the formula.
}
\author{
Ehouarn Le Faou
}
