% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{getRecords}
\alias{getRecords}
\title{Getting the simulation results}
\usage{
getRecords(metapop)
}
\arguments{
\item{metapop}{a \code{Metapopulation} objects}
}
\value{
A list where each item is associated with a simulation. Each of
these elements consists of a list of data.frames, one per population.
These data.frames consist of the same columns as the results
(see \link[Ease]{getResults} documentation), except that they do not
include the stop conditions.
}
\description{
Getting the simulation results
}
\author{
Ehouarn Le Faou
}
