% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFunctions.R
\name{IDgenotypeGeneration}
\alias{IDgenotypeGeneration}
\title{Genotype identifier}
\usage{
IDgenotypeGeneration(dl1, dl2, hl = NULL)
}
\arguments{
\item{dl1}{character (or factors) vector. The first diploid haplotype}

\item{dl2}{character (or factors) vector. the second diploid haplotype}

\item{hl}{character (or factors) vector. the haploid haplotype}
}
\value{
The genotype ID as a character string.
}
\description{
Generation of the input genotype ID, i.e. the concatenation in string form
of the names of the alleles constituting these haplotypes (the two from
the diploid genome and the one from the haploid genome).
}
\author{
Ehouarn Le Faou
}
